refactor: change code location and color gradiant of repair attempt graph
