+++ report-app/src/styles.scss
@@ -32,22 +32,18 @@
   --status-fill-poor: #ef4444;
   --status-fill-neutral: #aaa;
 
+  /* When we need a more gradiant spread of "meh". */
+  --status-fill-mediocre-1: #fbbc04; /* Yellow 500 */
+  --status-fill-mediocre-2: #f9ab00; /* Yellow 600 */
+  --status-fill-mediocre-3: #f29900; /* Yellow 700 */
+  --status-fill-mediocre-4: #ea8600; /* Yellow 800 */
+  --status-fill-mediocre-5: #e37400; /* Yellow 900 */
+
   --status-text-excellent: #0c855d;
   --status-text-great: #0c855d; // TODO: do we want to differentiate from `excellent`?
   --status-text-good: #c57f08;
   --status-text-poor: #eb1515;
   --status-text-neutral: #64748b;
-
-  /* 10-step Green-to-Red Quality Gradient */
-  --chart-grade-1: #10b981; /* Emerald 500 (Excellent) */
-  --chart-grade-2: #22c55e; /* Green 500 */
-  --chart-grade-3: #4ade80; /* Green 400 */
-  --chart-grade-4: #84cc16; /* Lime 500 (Great) */
-  --chart-grade-5: #a3e635; /* Lime 400 */
-  --chart-grade-6: #facc15; /* Yellow 400 */
-  --chart-grade-7: #f59e0b; /* Amber 500 (Good) */
-  --chart-grade-8: #f97316; /* Orange 500 */
-  --chart-grade-9: #ef4444; /* Red 500 (Poor) */
 
   --tooltip-background-color: light-dark(#111827, #f1f4f9);
   --tooltip-text-color: light-dark(#f9fafb, #1e293b);
